+++ packages/calendar-range/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.2.5](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-calendar-range@4.2.4...@alfalab/core-components-calendar-range@4.2.5) (2022-02-15)
+
+**Note:** Version bump only for package @alfalab/core-components-calendar-range
+
+
+
+
+
 ## [4.2.4](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-calendar-range@4.2.3...@alfalab/core-components-calendar-range@4.2.4) (2022-02-09)
 
 **Note:** Version bump only for package @alfalab/core-components-calendar-range